Migracion de unidades a Kilos

+++ src/main/java/com/pezfa/inventario/controller/CompraEspecieController.java
@@ -79,7 +79,7 @@ public class CompraEspecieController implements Serializable
         this.usuarioController = usuarioController;
     }
 
-    public float obtenerCostoTotal()
+    public double obtenerCostoTotal()
     {
         try
         {
